requirements.txt
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
ib-insync>=0.9.86
python-json-logger>=2.0
pytest>=8.0
httpx>=0.27
google-generativeai>=0.3.0
python-dotenv>=1.0.0